# corporate-repos/psc-sre-automacao-controller/src/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSC SRE Automacao Controller",
    "description": "API do Controller de automacao SRE. Orquestra a execucao de automacoes nos Agents distribuidos por cluster, gerencia logs de execucao, autenticacao JWT e integracao com o OAS (Orquestrador de Automacao de Servicos).",
    "version": "3.6.6",
    "contact": {
      "name": "PSC SRE Automacao"
    }
  },
  "servers": [
    {
      "url": "https://sre-aut-controller.psc.{cluster}.bb.com.br",
      "description": "Controller por cluster",
      "variables": {
        "cluster": {
          "description": "Nome do cluster OpenShift",
          "default": "k8shmlbb111b"
        }
      }
    }
  ],
...
  "paths": {
...
    "/auth/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Emitir token JWT",
        "description": "Emite um token JWT para autenticacao nas rotas protegidas. Requer API Key no header x-api-key. O token e assinado com HS256 e inclui os scopes solicitados.",
        "operationId": "issueToken",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subject": {
                    "type": "string",
                    "description": "Subject do token (sub claim)",
                    "example": "helmfire-ritmo-client"
                  },
                  "scope": {
                    "description": "Scope(s) solicitado(s). String separada por espaco/virgula ou array.",
                    "oneOf": [
                      {
                        "type": "string",
                        "example": "execute:automation send:logs"
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    ]
                  },
                  "expiresIn": {
                    "type": "string",
                    "description": "Tempo de expiracao do token (ex: 5m, 1h, 3600)",
                    "example": "5m"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token emitido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Scopes invalidos ou ausentes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API Key invalida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Scopes solicitados nao permitidos para esta API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT emitido por POST /auth/token. Algoritmo HS256, issuer=psc-sre-automacao-controller."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key para emissao de tokens JWT e consulta de scopes. Configurada via AUTH_API_KEYS_SCOPES no Kubernetes Secret."
      },
      "internalOrigin": {
        "type": "apiKey",
        "in": "header",
        "name": "x-techbb-namespace",
        "description": "Autenticacao por origem interna do Kubernetes. Quando os headers x-techbb-namespace e x-techbb-service-account correspondem ao caller confiavel configurado, a autenticacao JWT e dispensada."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Mensagem de erro"
          },
          "detail": {
            "type": "string",
            "description": "Detalhes adicionais do erro"
          },
          "execId": {
            "type": "string",
            "format": "uuid",
            "description": "ID da execucao, quando disponivel"
          }
        }
      },
...
      "TokenResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT assinado"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "string",
            "example": "5m"
          },
          "howToUse": {
            "type": "string",
            "example": "Authorization: Bearer <token>"
          }
        }
      },
...
    }
  }
}